Convert ItemCode from int to ulong to match it with NetworkObject PrefabHash

+++ Assets/Scripts/SceneObjects/ShoppingCart/ShoppingCartItem.cs
@@ -23,7 +23,7 @@ public class ShoppingCartItem : NetworkBehaviour
 
     private int _nextIndex = 0;
 
-    private int[] _itemCodes;
+    private ulong[] _itemCodes;
     private bool[] _occupiedPositions;
     private List<GameObject> _itemPositions;
 
@@ -35,7 +35,7 @@ public class ShoppingCartItem : NetworkBehaviour
     {
         // Items
         _itemPositions = gameObject.FindChildrenWithTag(ITEM_POSITIONS_TAG);
-        _itemCodes = new int[_itemPositions.Count];
+        _itemCodes = new ulong[_itemPositions.Count];
         _occupiedPositions = new bool[_itemPositions.Count];
 
         _ownerID.OnValueChanged += onOwnershipChanged;
@@ -118,7 +118,7 @@ public class ShoppingCartItem : NetworkBehaviour
         }
     }
 
-    private void setNextItem(int itemTypeCode)
+    private void setNextItem(ulong itemTypeCode)
     {
         // Destroy previous model
         _itemPositions[_nextIndex].transform.DestroyAllChildren();
@@ -173,7 +173,7 @@ public class ShoppingCartItem : NetworkBehaviour
             int colorNumber = NetworkManager.ConnectedClients[playerID].PlayerObject.GetComponent<PlayerInfo>().PlayerData.Color;
             UpdateCartColor(colorNumber);
             setCartColor_ClientRpc(colorNumber);
-            
+
             _ownerID.Value = playerID;
         }
     }
@@ -208,7 +208,7 @@ public class ShoppingCartItem : NetworkBehaviour
     }
 
     [ClientRpc]
-    private void setNextItem_ClientRpc(int itemTypeCode)
+    private void setNextItem_ClientRpc(ulong itemTypeCode)
     {
         if(!IsServer)
         {
